add image, @auth/core, merge adapter.ts to index.ts
